@@ -5,12 +5,13 @@
     "keywords": ["laravel", "framework"],
     "license": "MIT",
     "require": {
-        "php": "^8.2",
+        "php": "^8.3",
         "intervention/image": "^2",
         "laravel/framework": "^11.0",
         "laravel/sanctum": "^4.0",
         "laravel/tinker": "^2.9",
         "laravel/ui": "^4.5",
+        "maatwebsite/excel": "^3.1",
         "realrashid/sweet-alert": "^7.1",
         "spatie/laravel-html": "^3.9",
         "spatie/laravel-permission": "^6.7"
