@@ -19,6 +19,7 @@
                         <th>Name</th>
                         <th>No Seri</th>
                         <th class="text-center">Image</th>
+                        <th class="text-center">Detail</th>
                         <th class="text-center">Status</th>
                     </tr>
                     </thead>
@@ -27,11 +28,9 @@
                         @foreach($sale as $key => $data)
                             <td>{{$key +1}}</td>
                             <td>
-                                <a href="{{route('superadmin.item.show', $data->id)}}"
-                                   class="text-dark">{{$data->name}}</a>
+                                {{$data->item->name}}
                             </td>
-                            <td>{{ optional($data->cat)->name }} - {{ $data->no_seri }}
-                            </td>
+                            <td>{{$data->item->cat->name}}-{{$data->item->no_seri}}</td>
                             <td class="text-center">
                                 @if($data->image)
                                     <button data-bs-toggle="modal" data-bs-target="#exampleModal{{$data->id}}"
@@ -63,13 +62,77 @@
 
                             </td>
                             <td class="text-center">
+                                <div class="modal fade" id="exampleVerticallycenteredModal{{$data->id}}" tabindex="-1"
+                                     aria-hidden="true">
+                                    <div class="modal-dialog modal-dialog-centered">
+                                        <div class="modal-content">
+                                            <div class="modal-header">
+                                                <h5 class="modal-title">Descrpti Maintenance</h5>
+                                                <button type="button" class="btn-close" data-bs-dismiss="modal"
+                                                        aria-label="Close"></button>
+                                            </div>
+                                            <form action="{{route('mainten.store')}}"
+                                                  method="POST">
+                                                @csrf
+                                                <div class="modal-body">
+                                                    <input value="{{$data->id}}" type="hidden" name="item_id"
+                                                           class="form-control">
+                                                    <label class="col-form-label">Descript</label>
+                                                    <textarea type="text" name="descript"
+                                                              class="form-control"
+                                                              placeholder="Enter Descript"></textarea>
+                                                </div>
+                                                <div class="modal-footer">
+                                                    <button type="button" class="btn btn-secondary"
+                                                            data-bs-dismiss="modal">Close
+                                                    </button>
+                                                    <button type="submit" class="btn btn-primary">Save<i
+                                                            class="bx bx-save"></i></button>
+                                                </div>
+                                            </form>
+                                        </div>
+                                    </div>
+                                </div>
+                                <button data-bs-toggle="modal" data-bs-target="#exampleLargeModal{{$data->id}}"
+                                        class="btn btn-dnd btn-sm lni lni-eye me-1" title="view">
+                                </button>
+                                <div class="modal fade" id="exampleLargeModal{{$data->id}}" tabindex="-1"
+                                     aria-hidden="true">
+                                    <div class="modal-dialog modal-lg">
+                                        <div class="modal-content">
+                                            <div class="modal-header">
+                                                <h5 class="modal-title">Descript Sale</h5>
+                                                <button type="button" class="btn-close" data-bs-dismiss="modal"
+                                                        aria-label="Close"></button>
+                                            </div>
+                                            <div class="modal-body">
+                                                <div class="table-responsive">
+                                                    <table id="" class="table table-bordered">
+                                                        <tr>
+                                                            <th width="5%"><div class="float-start">Name</div></th>
+                                                            <td><div class="float-start">{{$data->item->name}}</div></td>
+                                                        </tr>
+                                                        <tr>
+                                                            <th><div class="float-start">No Seri</div></th>
+                                                            <td><div class="float-start">{{$data->item->no_seri}}</div></td>
+                                                        </tr>
+                                                        <tr>
+                                                            <th><div class="float-start">Descript</div></th>
+                                                            <td><div class="float-start">{{($data->descript)}}</div></td>
+                                                        </tr>
+                                                    </table>
+                                                </div>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                            </td>
+                            <td class="text-center">
                                 <span class="badge bg-warning">Sale</span>
                             </td>
-
                     </tr>
                     @endforeach
                     </tbody>
-                    </tfoot>
                 </table>
             </div>
         </div>
