@@ -12,10 +12,30 @@
     <!--navigation-->
     <ul class="metismenu" id="menu">
         <li>
-            <a href="{{url('/dashboard/')}}">
+            <a href="{{url('employe/dashboard/')}}">
                 <div class="parent-icon"><i class='bx bx-home-alt'></i>
                 </div>
                 <div class="menu-title">Dashboard</div>
+            </a>
+        </li>
+        <li>
+
+            <a href="{{route('employe.rental')}}">
+                <div class="parent-icon"><i class='bx bx-history'></i>
+                </div>
+                <div class="menu-title">Rental
+                    @if($rental > 0)
+                        <span class="badge bg-danger">{{$rental}}</span>
+                    @else
+                    @endif
+                </div>
+            </a
+        <li>
+
+            <a href="{{route('employe.rentaldivisi.index')}}">
+                <div class="parent-icon"><i class='lni lni-apartment'></i>
+                </div>
+                <div class="menu-title">Rental Divisi</div>
             </a>
         </li>
         <li class="menu-label">Items</li>
@@ -43,11 +63,24 @@
             </a>
         </li>
         <li>
-            <a href="{{route('employe.mainten')}}">
+            <a href="{{route('employe.accesSale.index')}}">
+                <div class="parent-icon"><i class='bx bx-dollar'></i>
+                </div>
+                <div class="menu-title">Accessories Sale</div>
+            </a>
+        </li>
+        <li>
+            <a href="javascript:;" class="has-arrow">
                 <div class="parent-icon"><i class='bx bx-shield-quarter'></i>
                 </div>
                 <div class="menu-title">Maintenance</div>
             </a>
+            <ul>
+                <li>
+                    <a href="{{route('employe.mainten')}}"><i class="bx bx-box"></i> Item</a>
+                    <a href="{{route('employe.mainten.access')}}"><i class="bx bx-collection"></i> Accessories</a>
+                </li>
+            </ul>
         </li>
     </ul>
     <!--end navigation-->
