@@ -19,28 +19,31 @@
             <div class="table-responsive">
             <table id="accessories" class="table table-striped table-bordered" style="width:100%">
                 <thead>
-                    <tr>
-                        <th width="2%">No</th>
-                        <th>Name</th>
-                        <th>Stok All</th>
-                        <th>Stok Available</th>
-                        <th>Rental</th>
-                        <th class="text-center" width="9%">Action</th>
-                    </tr>
+                <tr>
+                    <th width="2%">No</th>
+                    <th>Name</th>
+                    <th>Stok All</th>
+                    <th>Stok Available</th>
+                    <th>Rental</th>
+                    <th>Rental Divisi</th>
+                    <th>Maintenance</th>
+                    <th class="text-center" width="9%">Action</th>
+                </tr>
                 </thead>
                 <tbody>
-                    @foreach($accessoriesData as $key => $data)
+                @foreach($accessoriesData as $key => $data)
                     <tr>
                         <td data-index="{{ $key + 1 }}">{{ $key + 1 }}</td>
                         <td>{{ $data['name'] }}</td>
                         <td>{{ $data['stok_all'] }}</td>
                         <td>{{ $data['stok'] }}</td>
                         <td>{{ $data['rentedQty'] }}</td>
+                        <td>{{ $data['borrowedQty'] }}</td>
+                        <td>{{ $data['maintenanceQty'] }}</td>
                         <td>
                             <a href="{{ route('admin.acces.destroy', $data['id']) }}" data-confirm-delete="true" class="btn btn-danger btn-sm bx bx-trash" data-bs-toggle="tooltip" data-bs-placement="top" title="Delete"></a>
-                            <button data-bs-toggle="modal" data-bs-target="#exampleVerticallycenteredModal{{ $data['id'] }}" class="btn btn-warning btn-sm float-end bx bx-edit ms-2" data-bs-placement="top" title="Edit"></button>
 
-                            <!-- Edit Modal -->
+                            <button data-bs-toggle="modal" data-bs-target="#exampleVerticallycenteredModal{{ $data['id'] }}" class="btn btn-warning btn-sm float-end bx bx-edit ms-2" data-bs-placement="top" title="Edit"></button>
                             <div class="modal fade" id="exampleVerticallycenteredModal{{ $data['id'] }}" tabindex="-1" aria-hidden="true">
                                 <div class="modal-dialog modal-dialog-centered">
                                     <div class="modal-content">
@@ -67,10 +70,46 @@
                                     </div>
                                 </div>
                             </div>
-                            <!-- End of Edit Modal -->
+
+                            <button data-bs-toggle="modal"
+                                    data-bs-target="#tambahstok{{ $data['id'] }}"
+                                    class="btn btn-dnd btn-sm float-end bx bx-edit ms-2" data-bs-placement="top"
+                                    title="Edit"></button>
+                            <div class="modal fade" id="tambahstok{{ $data['id'] }}"
+                                 tabindex="-1" aria-hidden="true">
+                                <div class="modal-dialog modal-dialog-centered">
+                                    <div class="modal-content">
+                                        <div class="modal-header">
+                                            <h5 class="modal-title">Edit Accessories</h5>
+                                            <button type="button" class="btn-close" data-bs-dismiss="modal"
+                                                    aria-label="Close"></button>
+                                        </div>
+                                        <form action="{{ route('manager.acces.tambah', $data['id']) }}"
+                                              method="POST">
+                                            @csrf
+                                            @method('PUT')
+                                            <div class="modal-body">
+                                                <label class="col-form-label">Name Accessories</label>
+                                                <input value="{{ $data['name'] }}" type="text" name="name"
+                                                       class="form-control" placeholder="Enter Accessories" readonly>
+                                                <label class="col-form-label mt-2">qty</label>
+                                                <input type="number" value="0" name="stok"
+                                                       class="form-control" placeholder="">
+                                            </div>
+                                            <div class="modal-footer">
+                                                <button type="button" class="btn btn-secondary"
+                                                        data-bs-dismiss="modal">Close
+                                                </button>
+                                                <button type="submit" class="btn btn-primary">Save<i
+                                                        class="bx bx-save"></i></button>
+                                            </div>
+                                        </form>
+                                    </div>
+                                </div>
+                            </div>
                         </td>
                     </tr>
-                    @endforeach
+                @endforeach
                 </tbody>
             </table>
             </div>
@@ -106,7 +145,7 @@
                     </tr>
                     </thead>
                     <tbody>
-                    @foreach($rental as $key => $data)
+                    @foreach($rentals as $key => $data)
                         <tr>
                             <td data-index="{{ $key + 1 }}">{{ $key + 1 }}</td>
                             <td>
